requirements.txt
Beaker==1.11.0
bottle==0.12.18
bottle-beaker==0.1.3
bottle-cork==0.12.0
pycrypto==2.6.1
